[CO2 board-rev3] Add pin selection to Touch Sensor
Behind the front OLED display there now is a set of pads to select what GPIO could be used for the touch sensor.
This required a redesign of the traces on the front and the back of the PCB.
Now in general the 3.3v and GND lines are on the front and the I2C lines are on the back.